Major refactore, replacing hyper for reqwest, adding failure for error-handling, removing unnecessary modules
