+++ Cargo.toml
@@ -9,8 +9,10 @@ documentation = "http://jgillich.github.io/hyperdav/hyperdav/webdav/index.html"
 
 [dependencies]
 xml-rs = "0.3"
-hyper = "0.9"
+reqwest = "0.8"
 url = "1.1"
+failure = "0.1"
+hyper = "0.11"
 
 [dev-dependencies]
 uuid = "0.1"
